backend/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==11.3.0
